+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/overlay/CommandMenuGui.java
@@ -852,6 +852,7 @@ public class CommandMenuGui extends OverlayBase {
 			ClientUtils.CM_ELEMENT.endTransform(guiGraphics);
 		}
 		guiGraphics.flush();
+
 		guiGraphics.managed = false;
 	}